+++ web/packages/pyxis-components/src/public/molecules/PubShowRow/PubShowRow.tsx
@@ -1,4 +1,4 @@
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import type { Show } from 'pyxis-types';
 
 export type PubShowRowProps = {